requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.0
python-dotenv==1.0.0
pydantic[email]==2.5.3
python-multipart==0.0.6
Pillow==10.2.0
PyPDF2==3.0.1
google-generativeai==0.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
